+++ src/main/java/com/github/bluekey/dto/common/TotalAndGrowthDto.java
@@ -12,12 +12,12 @@ import lombok.RequiredArgsConstructor;
 @Schema(description = "총액과 증감률")
 public class TotalAndGrowthDto {
 	@Schema(description = "총액", example = "1000000")
-	private Long totalAmount;
+	private Integer totalAmount;
 	@Schema(description = "증감률", example = "2.1")
 	private Double growthRate;
 
 	@Builder
-	public TotalAndGrowthDto(final Long totalAmount, final Double growthRate) {
+	public TotalAndGrowthDto(final Integer totalAmount, final Double growthRate) {
 		this.totalAmount = totalAmount;
 		this.growthRate = growthRate;
 	}
